+++ lab1/src/core/table/hashtable.hpp
@@ -8,13 +8,13 @@ public:
     ~HashTable() override = default;
     
     bool add(const Record& rec) override;
-    size_t find(const std::string& key) override;
-    size_t getTotalCollisions() const;
+    int find(const std::string& key) override;
+    int getTotalCollisions() const;
 
 private:
     bool linear_probing(int& index, const std::string& key);
     int m_hashFunction(const std::string& key) const;
 
-    size_t m_totalCollisions {};
+    int m_totalCollisions {};
 
 };
