+++ lab1/src/core/table/itable.hpp
@@ -18,7 +18,7 @@ public:
     virtual bool modify(std::vector<std::string> fieldsNew) = 0;
     virtual bool add(const Record& rec) = 0;
     virtual bool remove(const std::string& key) = 0;
-    virtual size_t find(const std::string& key) = 0;
+    virtual int find(const std::string& key) = 0;
 
     virtual std::vector<Record> getData() const = 0;
     virtual std::vector<std::string> getTitles() const = 0;
