+++ lab1/src/core/table/regulartable.hpp
@@ -8,6 +8,6 @@ public:
     ~RegularTable() override = default;
 
     bool add(const Record& rec) override;
-    size_t find(const std::string& key) override;
+    int find(const std::string& key) override;
 
 };
